Convert hasRole to TS and fix/improve its usage
