+++ app/livechat/server/agentStatus.js
@@ -1,10 +1,10 @@
 import { UserPresenceMonitor } from 'meteor/konecty:user-presence';
 
 import { Livechat } from './lib/Livechat';
-import { hasRole } from '../../authorization';
+import { hasAnyRole } from '../../authorization/server/functions/hasRole';
 
 UserPresenceMonitor.onSetUserStatus((user, status) => {
-	if (hasRole(user._id, 'livechat-manager') || hasRole(user._id, 'livechat-monitor') || hasRole(user._id, 'livechat-agent')) {
+	if (hasAnyRole(user._id, ['livechat-manager', 'livechat-monitor', 'livechat-agent'])) {
 		Livechat.notifyAgentStatusChanged(user._id, status);
 	}
 });
