+++ ee/app/livechat-enterprise/server/methods/getUnitsFromUserRoles.ts
@@ -1,0 +1,15 @@
+import { Meteor } from 'meteor/meteor';
+
+import { hasRolesAsync } from '../../../../../app/authorization/server/functions/hasRole';
+import LivechatUnit from '../../../models/server/models/LivechatUnit';
+
+Meteor.methods({
+	async 'livechat:getUnitsFromUserRoles'() {
+		const user = Meteor.user();
+		if (!user || await hasRolesAsync(user._id, ['admin', 'livechat-manager'])) {
+			return;
+		}
+
+		return await hasRolesAsync(user._id, ['livechat-monitor']) && LivechatUnit.findByMonitorId(user._id);
+	},
+});
