+++ server/methods/OEmbedCacheCleanup.js
@@ -1,12 +1,12 @@
 import { Meteor } from 'meteor/meteor';
 
 import { OEmbedCache } from '../../app/models/server/raw';
-import { settings } from '../../app/settings';
-import { hasRole } from '../../app/authorization';
+import { settings } from '../../app/settings/server';
+import { hasRolesAsync } from '../../app/authorization/server/functions/hasRole';
 
 Meteor.methods({
 	async OEmbedCacheCleanup() {
-		if (Meteor.userId() && !hasRole(Meteor.userId(), 'admin')) {
+		if (Meteor.userId() && !await hasRolesAsync(Meteor.userId(), ['admin'])) {
 			throw new Meteor.Error('error-not-allowed', 'Not allowed', {
 				method: 'OEmbedCacheCleanup',
 			});
